TECHNEX2026BAD/requirements.txt
absl-py==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
certifi==2026.2.25
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.134.0
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.15.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.2
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.6.1
numpy==2.3.5
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==3.0.1
pillow==11.3.0
inference-sdk==1.0.2
polars==1.38.1
polars-runtime-32==1.38.1
protobuf==7.34.0
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==82.0.0
six==1.17.0
starlette==0.52.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
torch==2.10.0
torchvision==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.18
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.41.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.1
